pyproject.toml
[tool.poetry]
name = "maintenance-system"
version = "0.1.0"
description = "Its a project to maintain records for SilverFoxStudio"
authors = ["Tahir Siddique <tahirsiddiquevhr@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
asgiref = "3.7.2"
certifi = "2023.7.22"
cffi = "1.15.1"
charset-normalizer = "3.2.0"
colorama = "0.4.6"
cryptography = "41.0.3"
defusedxml = "0.7.1"
django = "4.2.5"
django-allauth = "0.56.1"
django-cors-headers = "4.2.0"
djangorestframework = "3.14.0"
exceptiongroup = "1.1.3"
idna = "3.4"
iniconfig = "2.0.0"
oauthlib = "3.2.2"
packaging = "23.1"
pillow = "10.0.0"
pluggy = "1.3.0"
pycparser = "2.21"
pyjwt = "2.8.0"
pytest = "7.4.2"
pytest-django = "4.5.2"
python3-openid = "3.2.0"
pytz = "2023.3.post1"
requests = "2.31.0"
requests-oauthlib = "1.3.1"
sqlparse = "0.4.4"
tomli = "2.0.1"
typing-extensions = "4.7.1"
tzdata = "2023.3"
urllib3 = "2.0.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
